Initial commit
This is commit has a working API service based on Rust Actix-Web with
a dependency on mongodb for the datalayer. There are a couple
test which need to be corrected and a dockerfile which needs to
be added is this is to be used on a local desktop via container
deployment.

Signed-off-by: Toure Dunnon <tdunnon@gmail.com>